feat: adicionar FAQ à página Private e refatorar estilos; atualizar fontes e tokens
