+++ src/messages/pt.json
@@ -135,5 +135,25 @@
         "subtitle": "A Terra da Arte, Arquitetura e Design Nórdicos."
       }
     }
+  },
+  "PrivatePage": {
+    "faq": {
+      "titlePrefix": "Perguntas frequentes sobre os",
+      "titleHighlight": "Private",
+      "items": {
+        "whatIs": {
+          "title": "O que é a Experiência Private?",
+          "description": "O Private é uma experiência totalmente personalizada. Você escolhe o destino, as datas, quem irá acompanhá-lo e o ritmo da viagem. A partir dos seus interesses, a equipe HÁRPÁ desenvolve uma jornada exclusiva, desenhada sob medida para você. Caso deseje, é possível utilizar todo o itinerário da Expedição em Grupo e adaptar às suas preferências."
+        },
+        "guide": {
+          "title": "O guia acompanha a viagem Private?",
+          "description": "Não necessariamente. Diferentemente das Expedições em Grupo, a modalidade Private não inclui acompanhamento presencial durante a viagem. Ainda assim, a equipe HÁRPÁ permanece disponível 24h por dia via WhatsApp para suporte, orientações e eventuais ajustes ao longo da jornada."
+        },
+        "whyChoose": {
+          "title": "Por que escolher a Experiência Private?",
+          "description": "A experiência Private é ideal para quem busca total liberdade para viajar no seu próprio ritmo. Com datas flexíveis, destinos à sua escolha e um roteiro desenvolvido exclusivamente para os seus interesses, cada jornada é acuradamente planejada para refletir aquilo que torna uma viagem verdadeiramente especial para você."
+        }
+      }
+    }
   }
 }
