+++ src/patterns/Faq/index.tsx
@@ -1,0 +1,35 @@
+import Details from '@/components/Details'
+import Container from '../Container'
+import * as S from './styles'
+
+type FaqItem = {
+  title: string
+  description: string[]
+}
+
+type Props = {
+  titlePrefix: string
+  titleHighlight: string
+  items: FaqItem[]
+}
+
+const Faq = ({ titlePrefix, titleHighlight, items }: Props) => (
+  <S.Faq>
+    <Container>
+      <S.FaqWrapper>
+        <S.FaqTitle>
+          {titlePrefix} <strong>{titleHighlight}</strong>
+        </S.FaqTitle>
+        {items.map((item, index) => (
+          <Details
+            key={index}
+            title={item.title}
+            description={item.description}
+          />
+        ))}
+      </S.FaqWrapper>
+    </Container>
+  </S.Faq>
+)
+
+export default Faq
